requirements.txt
langchain>=0.1.12
langchain-core>=0.1.27
langchain-openai>=0.0.8
langchain-chroma>=0.0.15
langchain-anthropic~=0.3.15
langgraph
chromadb>=0.4.24
openai>=1.12.0
python-dotenv>=1.0.1
boto3~=1.38.36
botocore~=1.38.36
SQLAlchemy~=2.0.41
alembic~=1.13.0
requests~=2.32.4
PyYAML~=6.0.2
toml~=0.10.2
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0
chainlit>=2.9.5
pandas
